Move the alt-table logic into a handler
